// src/store/posts/types.ts
enum ActionType {
    ADD_POST_REQUEST = "ADD_POST_REQUEST",
    ADD_POST_SUCCESS = "ADD_POST_SUCCESS",
    ADD_POST_ERROR = "ADD_POST_ERROR",
    EDIT_POST_REQUEST = "EDIT_POST_REQUEST",
    EDIT_POST_SUCCESS = "EDIT_POST_SUCCESS",
    EDIT_POST_ERROR = "EDIT_POST_ERROR",
    DELETE_POST_REQUEST = "DELETE_POST_REQUEST",
    DELETE_POST_SUCCESS = "DELETE_POST_SUCCESS",
    DELETE_POST_ERROR = "DELETE_POST_ERROR",
    FETCH_REQUEST = "FETCH_REQUEST",
    FETCH_SUCCESS = "FETCH_SUCCESS",
    FETCH_ERROR = "FETCH_ERROR",
    SET_LOADING = "SET_LOADING"
}

export default ActionType;
 